Pull Request comments are addressed
